Put fragments behind an extension
Fragments allow you to have a Ducktype file start with just sections or
block elements, instead a page heading. This is implemented, but it needs
to be tested more thoroughly, especially with includes. Consequently, it
didn't make it into the 1.0 spec. Since it's not in the 1.0 spec, we need
to make fragments an error. If you want to use them, use this:

@ducktype/1.0 __future__/fragments