Rework depth handling and add beginnings of NodeFactory
Some of the depth handling code was just wrong with respect to
section fragments. This cleans up depth handling and adds Division
as a subclass of Node. I've also added NodeFactory, which will be
responsible for all node creation. This will be necessary for
extensions that create other output formats.